refactor: 서비스 CRD로 분류하지 않고 하나로 통일
