+++ src/test/java/roomescape/service/theme/ThemeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service.reservationtime;
+package roomescape.service.theme;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,23 +11,23 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationTimeDeleteServiceTest {
+class ThemeServiceTest {
 
     @Autowired
-    private ReservationTimeDeleteService reservationTimeDeleteService;
+    private ThemeService themeService;
 
     @Test
-    @DisplayName("예약 중이 아닌 시간을 삭제할 시 성공한다.")
+    @DisplayName("예약 중이 아닌 테마를 삭제할 시 성공한다.")
     void deleteNotReservedTime_Success() {
-        assertThatCode(() -> reservationTimeDeleteService.deleteReservationTime(3L))
+        assertThatCode(() -> themeService.deleteTheme(3L))
                 .doesNotThrowAnyException();
     }
 
     @Test
-    @DisplayName("이미 예약 중인 시간을 삭제할 시 예외가 발생한다.")
+    @DisplayName("이미 예약 중인 테마를 삭제할 시 예외가 발생한다.")
     void deleteReservedTime_Failure() {
-        assertThatThrownBy(() -> reservationTimeDeleteService.deleteReservationTime(1L))
+        assertThatThrownBy(() -> themeService.deleteTheme(1L))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이미 예약중인 시간은 삭제할 수 없습니다.");
+                .hasMessage("이미 예약중인 테마는 삭제할 수 없습니다.");
     }
 }
